+++ libraries/matchers.rb
@@ -3,7 +3,7 @@
 # Cookbook Name:: iwork
 # Library:: matchers
 #
-# Copyright 2015 Jonathan Hartman
+# Copyright 2015-2016, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -19,11 +19,17 @@
 #
 
 if defined?(ChefSpec)
-  [:keynote_app, :numbers_app, :pages_app].each do |r|
-    ChefSpec.define_matcher(r)
+  {
+    keynote_app: %i(install upgrade),
+    numbers_app: %i(install upgrade),
+    pages_app: %i(install upgrade)
+  }.each do |matcher, actions|
+    ChefSpec.define_matcher(matcher)
 
-    define_method(:"install_#{r}") do |name|
-      ChefSpec::Matchers::ResourceMatcher.new(r, :install, name)
+    actions.each do |action|
+      define_method("#{action}_#{matcher}") do |name|
+        ChefSpec::Matchers::ResourceMatcher.new(matcher, action, name)
+      end
     end
   end
 end