+++ libraries/resource_numbers_app.rb
@@ -1,9 +1,9 @@
 # Encoding: UTF-8
 #
 # Cookbook Name:: iwork
-# Library:: numbers_app
+# Library:: resource_numbers_app
 #
-# Copyright 2015 Jonathan Hartman
+# Copyright 2015-2016, Jonathan Hartman
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -23,18 +23,18 @@ class Chef
     # A Chef resource for the Numbers app.
     #
     # @author Jonathan Hartman <j@p4nt5.com>
-    class NumbersApp < MacAppStoreApp
-      self.resource_name = :numbers_app
+    class NumbersApp < Resource
+      provides :numbers_app, platform_family: 'mac_os_x'
 
-      #
-      # Overload the app name with the one for this app.
-      #
-      attribute :app_name, kind_of: String, default: 'Numbers'
+      Chef::Resource::MacAppStoreApp.allowed_actions.each do |a|
+        action a do
+          include_recipe 'mac-app-store' unless a == :nothing
 
-      #
-      # Overload the bundle ID with the one for this app.
-      #
-      attribute :bundle_id, kind_of: String, default: 'com.apple.pkg.Numbers3'
+          mac_app_store_app 'Numbers' do
+            action a
+          end
+        end
+      end
     end
   end
 end