+++ spec/support/cookbooks/resource_numbers_app_test/recipes/default.rb
@@ -1,0 +1,7 @@
+# Encoding: UTF-8
+
+attrs = node['resource_numbers_app_test']
+
+numbers_app attrs['name'] do
+  action attrs['action'] unless attrs['action'].nil?
+end