+++ spec/support/cookbooks/resource_pages_app_test/recipes/default.rb
@@ -1,0 +1,7 @@
+# Encoding: UTF-8
+
+attrs = node['resource_pages_app_test']
+
+pages_app attrs['name'] do
+  action attrs['action'] unless attrs['action'].nil?
+end