+++ CHANGELOG.md
@@ -1,8 +1,9 @@
 Iwork Cookbook CHANGELOG
 ========================
 
-v?.?.? (????-??-??)
--------------------
+Unreleased
+----------
+- Update to use mac-app-store cookbook 2.0
 
 v0.1.0 (2015-05-10)
 -------------------